requirements.txt
streamlit>=1.47.0
openai>=1.97.0
selenium>=4.34.0
undetected-chromedriver>=3.5.0
beautifulsoup4>=4.13.0
python-dotenv>=1.1.0
requests>=2.32.0
browser-use>=0.5.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.177.0
httpx>=0.28.0 